Split custom queries unit tests
